Build only a single library type, static or shared
Typically, a static library should be more useful for UCDN. Use the
standard BUILD_SHARED_LIBS toggle of CMake to switch between the
library types, and link a static library by default.

If both library types are desired, they have to be built and installed
separately, but this should be a non-issue.