Add support for UAX#14 linebreak classes
Extract the linebreak class for each code point from the Unicode text
files and embed them in the database. Add a function for retrieving the
linebreak class given a code point. Add a second function resolving the
linebreak class according to the fallback outlined in LB1 of UAX#14.

This takes UAX#14 revision 35 into account as revised for Unicode 8.0.0.